# Make tests (excluding RegressionTests.jl) run 37% faster