--- a/test/runtests.jl
+++ b/test/runtests.jl
@@ -519,7 +519,8 @@ using Random: rand!
 
     @testset "Aqua" begin
         import Aqua
-        Aqua.test_all(Chairmarks, deps_compat=false)
+        # persistent_tasks=false because that test is slow and we don't use persistent tasks
+        Aqua.test_all(Chairmarks, deps_compat=false, persistent_tasks=false)
         Aqua.test_deps_compat(Chairmarks, check_extras=false)
     end