- **full pipeline seconds=.001 (29.3=>22.9)**
- **skip Aqua's persistent_task tests (22.9s => 20.5s)**
- **assorted other changes (reducing runtime budgets) (20.5s => 18.3s)**