@LilithHafner 2024-11-24
This brings us down to almost exactly the same test runtime as 1.2.2 (18.4s)