@LilithHafner 2024-11-24
CI times master => pr:

3m => 3m
5m => 5m
7m => 7m
15m => ...
15m => ...
4m => 3m
25m => ...
19m => ...
4m => 4m
17m => ...
17m => ...
5m => 4m
20m => ...
16m => ...
5m => 4m
3m => 3m
3s => 3s